% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{candidate_dopts}
\alias{candidate_dopts}
\title{Candidate dopt for direct search algorithms}
\usage{
candidate_dopts(dopt.SL.library, W, W_for_g, V, A, Y, newW, newV, newA,
  newY, QAW.reg, gAW, moMain_model = NULL, moCont_model = NULL,
  family = family)
}
\arguments{
\item{dopt.SL.library}{dopt library}

\item{W}{covariates}

\item{W_for_g}{covariates for estimating g}

\item{V}{subset of covariates for designing ODTR}

\item{A}{txt}

\item{Y}{outcome}

\item{newW}{new W}

\item{newV}{new V}

\item{newA}{new A}

\item{QAW.reg}{regression object for QAW}

\item{gAW}{P(A|W)}

\item{moMain_model}{for DynTxRegime}

\item{moCont_model}{for DynTxRegime}

\item{family}{family}

\item{new}{Y}
}
\value{

}
\description{
Candidate dopt predictionsfor direct search algorithms
}
