% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP.rc.discreteW}
\alias{DGP.rc.discreteW}
\title{DGP.rc.discreteW}
\usage{
DGP.rc.discreteW(n, a = NULL, kappa = NULL, QAW.fun)
}
\arguments{
\item{n}{n}

\item{a}{static txt}

\item{kappa}{prop can be treated in population}

\item{QAW.fun}{QAW function}
}
\value{
data for DGP.rc.discreteW
}
\description{
Generate data according to DGP RC - discrete W
}
