% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_AL_RC}
\alias{DGP_AL_RC}
\title{Simulate with AL bin DGP RC}
\usage{
DGP_AL_RC(n, dA = NULL, a = NULL, kappa = NULL, QAW.fun = QAW_AL_RC)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}

\item{kappa}{kappa}

\item{QAW.fun}{QAW.fun}
}
\value{
data for DGP_AL_RC
}
\description{
Generate data according to AL bin DGP RC
}
