% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_bin6}
\alias{DGP_bin6}
\title{Simulate with AL bin DGP6 with treat all optimal}
\usage{
DGP_bin6(n, dA = NULL, a = NULL, kappa = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}

\item{kappa}{for resource constraints}
}
\description{
Generate data according to AL bin DGP6 with treat all optimal
}
