% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_bin_complex}
\alias{DGP_bin_complex}
\title{Simulate with AL bin DGP}
\usage{
DGP_bin_complex(n, dA = NULL, a = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}
}
\value{
data for DGP_bin_complex
}
\description{
Generate data according to AL bin DGP
}
