% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_bin_complex_3tx}
\alias{DGP_bin_complex_3tx}
\title{Simulate with AL bin DGP}
\usage{
DGP_bin_complex_3tx(n, dA = NULL, a = NULL, kappa = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}

\item{kappa}{for resource constraints}
}
\value{

}
\description{
Generate data according to AL bin DGP
}
