% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_bin_simple}
\alias{DGP_bin_simple}
\title{Simulate with AL bin DGP with influential variable}
\usage{
DGP_bin_simple(n, dA = NULL, a = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}
}
\value{
data for DGP_bin_simple
}
\description{
Generate data according to AL bin DGP with influential variable
}
