% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_cont}
\alias{DGP_cont}
\title{Simulate with DB DGP}
\usage{
DGP_cont(n, dA = NULL, a = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}
}
\value{
data for DGP_cont
}
\description{
Generate data according to DB DGP
}
