% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_cont2}
\alias{DGP_cont2}
\title{Simulate with DB DGP with influential variable}
\usage{
DGP_cont2(n, dA = NULL, a = NULL, kappa = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}

\item{kappa}{for resource constraints}
}
\value{

}
\description{
Generate data according to DB DGP with influential variable
}
