% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_eff}
\alias{DGP_eff}
\title{Simulate with eff}
\usage{
DGP_eff(n, dA = NULL, a = NULL, kappa = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}

\item{kappa}{for resource constraints}
}
\description{
Generate data according to eff
}
