% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_eff_obs}
\alias{DGP_eff_obs}
\title{Simulate with eff}
\usage{
DGP_eff_obs(n, dA = NULL, a = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}
}
\value{
data for DGP_eff_obs
}
\description{
Generate data according to eff with obs g
}
