% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{DGP_null}
\alias{DGP_null}
\title{Simulate with null}
\usage{
DGP_null(n, dA = NULL, a = NULL)
}
\arguments{
\item{n}{n}

\item{dA}{rule type}

\item{a}{static txt}
}
\value{
data for DGP_null
}
\description{
Generate data according to null
}
