% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{estimatorsEYd_nonCVTMLE}
\alias{estimatorsEYd_nonCVTMLE}
\title{Estimators of E[Yd] that are not CV-TMLE}
\usage{
estimatorsEYd_nonCVTMLE(W, A, Y, d, QAW.reg, g.reg, ab, contrast)
}
\arguments{
\item{W}{covariates}

\item{A}{treatment}

\item{Y}{outcome}

\item{d}{txt under rule}

\item{QAW.reg}{Q(A,W) regression object}

\item{ab}{range of Y}

\item{contrast}{contrast}

\item{gAW.reg}{g(A|W) regression object}
}
\value{

}
\description{
Estimators of E[Yd] that are not CV-TMLE
}
