% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{estimatorsEYdopt_nonCVTMLE}
\alias{estimatorsEYdopt_nonCVTMLE}
\title{Estimators of E[Yd] that are not CV-TMLE}
\usage{
estimatorsEYdopt_nonCVTMLE(W, A, Y, dopt, QAW.reg, gAW, QAW.SL.library, ab)
}
\arguments{
\item{W}{covariates}

\item{A}{treatment}

\item{Y}{outcome}

\item{dopt}{txt under rule}

\item{QAW.reg}{QAW regression object}

\item{gAW}{P(A|W)}

\item{QAW.SL.library}{library for QAW}

\item{ab}{range of Y}
}
\value{

}
\description{
Estimators of E[Yd] that are not CV-TMLE
}
