% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{estimatorsEYgRC_nonCVTMLE}
\alias{estimatorsEYgRC_nonCVTMLE}
\title{Estimators of E[YgRC] that are not CV-TMLE}
\usage{
estimatorsEYgRC_nonCVTMLE(W, A, Y, rc.out, kappa, QAW.reg, g.reg, ab, contrast)
}
\arguments{
\item{W}{covariates}

\item{A}{treatment}

\item{Y}{outcome}

\item{rc.out}{rc.out}

\item{kappa}{kappa}

\item{QAW.reg}{Q(A,W) regression object}

\item{ab}{range of Y}

\item{contrast}{contrast}

\item{gAW.reg}{g(A|W) regression object}
}
\value{
psi and CI for non CVTMLE estimates
}
\description{
Estimators of E[YgRC] that are not CV-TMLE
}
