% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{estimatorsEYgstar_nonCVTMLE}
\alias{estimatorsEYgstar_nonCVTMLE}
\title{Estimators of E[Ygstar] that are not CV-TMLE}
\usage{
estimatorsEYgstar_nonCVTMLE(
  W,
  A,
  Y,
  gstar1W,
  gstar0W,
  QAW.reg,
  g.reg,
  ab,
  contrast
)
}
\arguments{
\item{W}{covariates}

\item{A}{treatment}

\item{Y}{outcome}

\item{gstar1W}{gstar1W}

\item{gstar0W}{gstar0W}

\item{QAW.reg}{Q(A,W) regression object}

\item{ab}{range of Y}

\item{contrast}{contrast}

\item{gAW.reg}{g(A|W) regression object}
}
\value{

}
\description{
Estimators of E[Ygstar] that are not CV-TMLE
}
