% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{getpreds.fun}
\alias{getpreds.fun}
\title{Predictions}
\usage{
getpreds.fun(Y, X, newX = NULL, family = gaussian(), SL.library,
  id = NULL, verbose = FALSE, control = list(), obsWeights = NULL,
  env = parent.frame())
}
\arguments{
\item{Y}{outcome}

\item{X}{predictors}

\item{newX}{new X}

\item{family}{family}

\item{SL.library}{SL library}

\item{id}{for unique ids}

\item{verbose}{verbose}

\item{control}{control}

\item{obsWeights}{obsWeights}

\item{env}{env}
}
\value{

}
\description{
Get library predictions (what would be library.predict). Code mostly from SuperLearner.
}
