% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6performanceFunctions.R
\name{performance_EYdopt}
\alias{performance_EYdopt}
\title{Performance EYdopt}
\usage{
performance_EYdopt(x, n, DGP_fun, QAW, QAW.SL.library, blip.SL.library,
  VFolds)
}
\arguments{
\item{x}{dummy}

\item{n}{n}

\item{DGP_fun}{DGP_fun}

\item{QAW}{QAW}

\item{QAW.SL.library}{library for QAW}

\item{blip.SL.library}{library for blip}

\item{VFolds}{CV Folds}
}
\value{

}
\description{
performance function for EYdopt
}
