% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6performanceFunctions.R
\name{performance_EYgRC}
\alias{performance_EYgRC}
\title{Performance EYgRC}
\usage{
performance_EYgRC(
  x,
  n,
  kappa,
  DGP_fun,
  QAW.fun,
  QAW.SL.library,
  blip.SL.library,
  grid.size,
  contrast = NULL
)
}
\arguments{
\item{x}{dummy}

\item{kappa}{risk.type}
}
\value{
performance metrics EYgRC
}
\description{
performance function for EYgRC
}
