% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6performanceFunctions.R
\name{performance_ODTR}
\alias{performance_ODTR}
\title{Performance ODTR}
\usage{
performance_ODTR(x, n, risk.type, DGP_fun, QAW, QAW.SL.library,
  blip.SL.library, dopt.SL.library, metalearner, grid.size, W_for_g = 1)
}
\arguments{
\item{x}{dummy}

\item{n}{n}

\item{risk.type}{risk.type}

\item{DGP_fun}{DGP_fun}

\item{QAW}{QAW}

\item{QAW.SL.library}{library for QAW}

\item{blip.SL.library}{library for blip}

\item{dopt.SL.library}{library for dopt}

\item{metalearner}{SL type}

\item{grid.size}{grid size}
}
\value{

}
\description{
performance function for the ODTR
}
