% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4odtr.R
\name{predict_odtr}
\alias{predict_odtr}
\title{Predict ODTR}
\usage{
predict_odtr(odtr_obj, newV)
}
\arguments{
\item{odtr_obj}{ODTR object from odtr function}

\item{newV}{new covariates to predict from}
}
\value{

}
\description{
Function that, given an odtr object, predicts the treatment under the ODTR for a new V.
}
\references{
Luedtke, Alexander R., and Mark J. van der Laan. "Super-learning of an optimal dynamic treatment rule." \emph{The international journal of biostatistics} 12.1 (2016): 305-332.
Coyle, J.R. (2017). Jeremy Coyle, “Computational Considerations for Targeted Learning” PhD diss., University of California, Berkeley 2017 \url{https://escholarship.org/uc/item/9kh0b9vm}.
Eric Polley, Erin LeDell, Chris Kennedy and Mark van der Laan (2018). SuperLearner: Super Learner Prediction. R package version 2.0-24. \url{https://CRAN.R-project.org/package=SuperLearner}.
}
