% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW.rc.allpos}
\alias{QAW.rc.allpos}
\title{QAW RC all pos}
\usage{
QAW.rc.allpos(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{
conditional mean of Y given A and W
}
\description{
Generate QAW according to DGP RC - all positive
}
