% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_AL_RC}
\alias{QAW_AL_RC}
\title{Simulate with AL bin DGP RC}
\usage{
QAW_AL_RC(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{
conditional mean of Y given A and W
}
\description{
Generate QAW according to AL bin DGP RC
}
