% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_bin2}
\alias{QAW_bin2}
\title{Simulate with AL bin DGP with positivity}
\usage{
QAW_bin2(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to AL bin DGP with positivity
}
