% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_bin3}
\alias{QAW_bin3}
\title{Simulate with AL bin DGP with influential variable}
\usage{
QAW_bin3(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to AL bin DGP with influential variable
}
