% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_bin4}
\alias{QAW_bin4}
\title{Simulate with AL bin DGP4}
\usage{
QAW_bin4(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to AL bin DGP4
}
