% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_bin5}
\alias{QAW_bin5}
\title{Simulate with AL bin DGP treat all optimal}
\usage{
QAW_bin5(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to AL bin DGP treat all optimal
}
