% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_bin_complex_3tx}
\alias{QAW_bin_complex_3tx}
\title{Simulate with AL bin DGP}
\usage{
QAW_bin_complex_3tx(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\description{
Generate QAW according to AL bin DGP
}
