% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_bin_dep}
\alias{QAW_bin_dep}
\title{Simulate with AL bin DGP dep W}
\usage{
QAW_bin_dep(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to AL bin dep W
}
