% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_bin_simple}
\alias{QAW_bin_simple}
\title{Simulate with AL bin DGP simple}
\usage{
QAW_bin_simple(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\description{
Generate QAW according to AL bin simple
}
