% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_cont}
\alias{QAW_cont}
\title{Simulate with DB DGP}
\usage{
QAW_cont(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{
conditional mean of Y given A and W
}
\description{
Generate QAW according to DB DGP
}
