% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_cont3}
\alias{QAW_cont3}
\title{Simulate with DB DGP with simple influential variable}
\usage{
QAW_cont3(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to DB DGP with simple influential variable
}
