% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_eff}
\alias{QAW_eff}
\title{Simulate with eff}
\usage{
QAW_eff(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\description{
Generate QAW according to eff
}
