% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_null}
\alias{QAW_null}
\title{Simulate with null}
\usage{
QAW_null(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to null
}
