% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DGPfunctions.R
\name{QAW_smooth2}
\alias{QAW_smooth2}
\title{Simulate with DB DGP}
\usage{
QAW_smooth2(A, W)
}
\arguments{
\item{A}{Vector of treatment}

\item{W}{Data frame of observed baseline covariates}
}
\value{

}
\description{
Generate QAW according to DB DGP
}
