% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7SLwrappers.R
\name{SL.blip.incorrect1}
\alias{SL.blip.incorrect1}
\title{Epi HTE GLM for blip 1}
\usage{
SL.blip.incorrect1(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\arguments{
\item{Y}{outcome}

\item{X}{predictors}

\item{newX}{new X}

\item{family}{family}

\item{obsWeights}{obsWeights}

\item{model}{model}

\item{...}{other}

\item{SL.library}{SL library}
}
\value{

}
\description{
Standard HTE GLM used in Epi analyses for blip 1.
}
