% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7SLwrappers.R
\name{SL.blip.W2}
\alias{SL.blip.W2}
\title{Blip W2}
\usage{
SL.blip.W2(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\arguments{
\item{Y}{outcome}

\item{X}{predictors}

\item{newX}{new X}

\item{family}{family}

\item{obsWeights}{obsWeights}

\item{model}{model}

\item{...}{other}

\item{SL.library}{SL library}
}
\description{
Blip W2.
}
