% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7SLwrappers.R
\name{SL.hal9001}
\alias{SL.hal9001}
\title{HAL with binning}
\usage{
SL.hal9001(
  Y,
  X,
  newX = NULL,
  max_degree = 2,
  fit_type = c("glmnet", "lassi"),
  n_folds = 10,
  use_min = TRUE,
  family,
  obsWeights = rep(1, length(Y)),
  bins = 75,
  ...
)
}
\arguments{
\item{Y}{Y}

\item{X}{X}

\item{newX}{newX}

\item{max_degree}{max_degree}

\item{fit_type}{fit_type}

\item{n_folds}{n_folds}

\item{use_min}{use_min}

\item{family}{family}

\item{obsWeights}{obsWeights}

\item{bins}{bins}
}
\value{

}
\description{
HAL with binning
}
