% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{SL.QAW.correct_smooth}
\alias{SL.QAW.correct_smooth}
\title{Correctly specified model for DB smooth function}
\usage{
SL.QAW.correct_smooth(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\arguments{
\item{Y}{outcome}

\item{X}{predictors}

\item{newX}{new X}

\item{family}{family}

\item{obsWeights}{obsWeights}

\item{model}{model}

\item{...}{other}

\item{SL.library}{SL library}
}
\value{

}
\description{
Correctly specified outcome regression model for DB smooth function
}
