% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7SLwrappers.R
\name{SL.QAW.HTEepi}
\alias{SL.QAW.HTEepi}
\title{Epi HTE GLM}
\usage{
SL.QAW.HTEepi(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\arguments{
\item{Y}{outcome}

\item{X}{predictors}

\item{newX}{new X}

\item{family}{family}

\item{obsWeights}{obsWeights}

\item{model}{model}

\item{...}{other}

\item{SL.library}{SL library}
}
\description{
Standard HTE GLM used in Epi analyses.
}
