% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7SLwrappers.R
\name{SL.QAW.W2}
\alias{SL.QAW.W2}
\title{QAW for W2}
\usage{
SL.QAW.W2(Y, X, newX, family, obsWeights, model = TRUE, ...)
}
\arguments{
\item{Y}{outcome}

\item{X}{predictors}

\item{newX}{new X}

\item{family}{family}

\item{obsWeights}{obsWeights}

\item{model}{model}

\item{...}{other}

\item{SL.library}{SL library}
}
\value{
SL.QAW.W2
}
\description{
QAW for W2.
}
