% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{tmle.fun}
\alias{tmle.fun}
\title{TMLE function}
\usage{
tmle.fun(A, Y, d, Qd, gAW, ab)
}
\arguments{
\item{A}{txt}

\item{Y}{outcome}

\item{d}{txt under rule}

\item{Qd}{outcome regression predictions under rule}

\item{gAW}{probability observed txt predictions}

\item{ab}{range of Y}
}
\value{

}
\description{
Compute E[Yd] using TMLE
}
