% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{tmle.g.fun}
\alias{tmle.g.fun}
\title{TMLE function}
\usage{
tmle.g.fun(A, Y, gstarAW, gstar1W, gstar0W, gAW, QAW, Q1W, Q0W, ab)
}
\arguments{
\item{A}{txt}

\item{Y}{outcome}

\item{gstarAW}{stochastic rule under A = A}

\item{gstar1W}{stochastic rule for A = 1}

\item{gstar0W}{stochastic rule for A = 0}

\item{gAW}{probability observed txt predictions}

\item{QAW}{outcome regression A = A}

\item{Q1W}{outcome regression A = 1}

\item{Q0W}{outcome regression A = 0}

\item{ab}{range of Y}
}
\description{
Compute E[Yg] using TMLE
}
