% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{tmle.rc.fun}
\alias{tmle.rc.fun}
\title{TMLE function}
\usage{
tmle.rc.fun(
  A,
  Y,
  gstarAW,
  gstar1W,
  gstar0W,
  gAW,
  QAW,
  Q1W,
  Q0W,
  ab,
  tauP,
  kappa
)
}
\arguments{
\item{A}{txt}

\item{Y}{outcome}

\item{gstarAW}{Prd.is.A}

\item{gstar1W}{Prd.is.1}

\item{gstar0W}{Prd.is.0}

\item{gAW}{probability observed txt predictions}

\item{QAW}{outcome regression A = A}

\item{Q1W}{outcome regression A = 1}

\item{Q0W}{outcome regression A = 0}

\item{ab}{range of Y}

\item{tauP}{tauP}

\item{kappa}{kappa}
}
\value{
psi and IC of tmle rc
}
\description{
Compute E[YgRC] using TMLE
}
