% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{dopt.fun}
\alias{dopt.fun}
\title{dopt function}
\usage{
dopt.fun(blip, kappa)
}
\arguments{
\item{blip}{predicted blip}

\item{kappa}{proportion of people who can be treated in population}
}
\description{
Compute dopt given blip
}
