% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{getpreds.dopt.fun}
\alias{getpreds.dopt.fun}
\title{Candidate dopt for direct search algorithms}
\usage{
getpreds.dopt.fun(dopt.SL.library, blip.SL.library = NULL, W, gform, V,
  A, Y, newV, QAW.reg, gAW, family = family)
}
\arguments{
\item{dopt.SL.library}{dopt library}

\item{W}{covariates}

\item{gform}{gform}

\item{V}{subset of covariates for designing ODTR}

\item{A}{txt}

\item{Y}{outcome}

\item{newV}{new V}

\item{QAW.reg}{regression object for QAW}

\item{gAW}{P(A|W)}

\item{family}{family}
}
\value{

}
\description{
Candidate dopt predictionsfor direct search algorithms
}
