% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2helperFunctions.R
\name{getpreds.dopt.fun}
\alias{getpreds.dopt.fun}
\title{Candidate dopt for direct search algorithms}
\usage{
getpreds.dopt.fun(
  dopt.SL.library,
  blip.SL.library = NULL,
  W,
  V,
  A,
  Y,
  newV,
  QAW.reg,
  g.reg,
  family = family
)
}
\arguments{
\item{dopt.SL.library}{dopt library}

\item{W}{covariates}

\item{V}{subset of covariates for designing ODTR}

\item{A}{txt}

\item{Y}{outcome}

\item{newV}{new V}

\item{QAW.reg}{regression object for Q(A,W)}

\item{g.reg}{regression object for g(A|W)}

\item{family}{family}

\item{gform}{gform}
}
\description{
Candidate dopt predictionsfor direct search algorithms
}
