% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6performanceFunctions.R
\name{performance_EYgstar}
\alias{performance_EYgstar}
\title{Performance EYgstar}
\usage{
performance_EYgstar(
  x,
  n,
  risk.type,
  cs_to_try,
  alphas_to_try,
  QAW.SL.library,
  blip.SL.library,
  DGP_fun,
  QAW.fun
)
}
\arguments{
\item{x}{dummy}

\item{risk.type}{risk.type}

\item{cs_to_try}{cs_to_try}

\item{alphas_to_try}{alphas_to_try}

\item{VFolds}{CV Folds}
}
\value{

}
\description{
performance function for EYgstar
}
