% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3SLodtr.R
\name{SL.blip}
\alias{SL.blip}
\title{Blip-based SL}
\usage{
SL.blip(V, W, A, Y, ab, QAW.reg, gAW, blip.SL.library, risk.type,
  grid.size, newV = NULL, VFolds)
}
\arguments{
\item{V}{subset covariates for designing ODTR}

\item{W}{covariates}

\item{A}{txt}

\item{Y}{outcome}

\item{ab}{range of Y}

\item{QAW.reg}{QAW regression object}

\item{gAW}{P(A|W)}

\item{blip.SL.library}{blip SL library}

\item{risk.type}{risk type}

\item{VFolds}{number of folds}

\item{grid}{size}

\item{new}{V}
}
\value{

}
\description{
SuperLearner for ODTR that returns estimated blips.
}
